+++ spec/requests/translate_spec.rb
@@ -266,6 +266,19 @@ describe DeepL::Requests::Translate do
         expect(request.options[:custom_instructions]).to eq(['Use informal language', 'Be concise'])
       end
     end
+
+    context 'when passing additional headers' do
+      it 'merges the headers into the request headers' do
+        request = described_class.new(api, nil, nil, nil, {},
+                                      { 'X-DeepL-Reporting-Tag' => 'my-tag' })
+        expect(request.send(:headers)).to include('X-DeepL-Reporting-Tag' => 'my-tag')
+      end
+
+      it 'defaults to no additional headers' do
+        request = described_class.new(api, nil, nil, nil)
+        expect(request.send(:headers).keys).to contain_exactly('Authorization', 'User-Agent')
+      end
+    end
   end
 
   describe '#request' do